+++ src/screensaver.ts
@@ -1,0 +1,11 @@
+export { screensaverActiveSignal } from "./features/shared/screensaverState.js";
+
+export {
+  AttractorSceneFeature,
+  type AttractorSceneFeatureOptions,
+} from "./features/attractor-scene/AttractorSceneFeature.js";
+
+export {
+  ScreensaverFeature,
+  type ScreensaverFeatureOptions,
+} from "./features/screensaver/ScreensaverFeature.js";